+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Recursion in `openDatabase` when using `SentrySqfliteDatabaseFactory` ([#3231](https://github.com/getsentry/sentry-dart/pull/3231))
 
+### Enhancements
+
+- Replay: continue processing if encountering `InheritedWidget` ([#3200](https://github.com/getsentry/sentry-dart/pull/3200))
+  - Prevents false debug warnings when using [provider](https://pub.dev/packages/provider) for example which extensively uses `InheritedWidget`
+
 ## 9.7.0-beta.2
 
 ### Features